requirements.txt
altgraph==0.17.4
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.8.3
cffi==1.17.1
channels==4.3.1
charset-normalizer==3.4.2
click==8.2.1
constantly==23.10.4
cryptography==45.0.6
daphne==4.2.1
Django==5.2.5
django-cors-headers==4.7.0
django-restframework==0.0.1
djangorestframework==3.16.1
h11==0.16.0
httptools==0.6.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
packaging==25.0
psutil==7.0.0
psutils==3.3.11
psycopg2-binary==2.9.10
puremagic==1.30
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
pyOpenSSL==25.1.0
pypdf==5.9.0
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
service-identity==24.2.0
setuptools==80.9.0
sniffio==1.3.1
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
zope.interface==7.2
